+++ src/api/user.ts
@@ -93,69 +93,3 @@ export const updateUser = async (userToUpdate: UserWithId, userId: string) => {
     throw new Error(`Failed to update user: ${error.response.data.message}`);
   }
 };
-
-/**
- * Get a user's balances
- * 🛑 CURRENTLY THIS ENDPOINT DOES NOT WORK 🛑
- */
-export const getBalances = async (userId: string) => {
-  try {
-    const res = await axios.get(
-      `${API_URL}/users/${userId}/balances`,
-      await withAuthHeader()
-    );
-    return res;
-  } catch (error: any) {
-    throw new Error(`Failed to get balances: ${error.response.data.message}`);
-  }
-};
-
-export const getOverview = async (address: string, last_days = 30) => {
-  try {
-    const res = await axios.get(
-      `${API_URL}/transactions/overview?account=${address}&last_days=${last_days}`,
-      await withAuthHeader()
-    );
-    return res;
-  } catch (error: any) {
-    throw new Error(`Failed to get overview`);
-  }
-};
-
-/**
- * Get a user's transactions
- * 🛑 CURRENTLY THIS ENDPOINT DOES NOT WORK 🛑
- */
-export const getTransactions = async (
-  address: string,
-  cursor = '',
-  limit = 20,
-  order = 'DESC',
-  from_date?: string,
-  to_date?: string
-) => {
-  const queryParams = new URLSearchParams({
-    account: address,
-    cursor: cursor.toString(),
-    limit: limit.toString(),
-    order,
-  });
-
-  if (from_date) {
-    queryParams.append('from_date', from_date);
-  }
-
-  if (to_date) {
-    queryParams.append('to_date', to_date);
-  }
-
-  try {
-    const res = await axios.get(
-      `${API_URL}/transactions?${queryParams.toString()}`,
-      await withAuthHeader()
-    );
-    return res;
-  } catch (error: any) {
-    throw new Error(`Failed to get transactions`);
-  }
-};
